Add context menu for clearing selected search terms, or all.
Would be a good idea to add a message box to ask them if they're sure they want to clear all search terms , and allow them to not have the message ask them again if they don't want it to show up again. This message should show up for both the context menu as well as the button, and the checkbox to not ask again will change a different setting depending on how they got to it. There should be a short message telling them how they got to the message, as well as a property that can be set that specifies which setting to modify if the user says to not ask them again (there will be one setting for if they click the button, and another setting if they used the context menu).