Only turn off column autosizing if marking an excessive...
number of rows at once so it doesn't look weird when only marking a few packages. The number of excessive rows might be changed, and it might be a good idea to allow the user to change it to whatever they want the number to be.

Also comment out turning off column autosizing when loading the package list as that may or may not help.