Temporary error handlers in case we can't find the source...
URL, the zip file, or the manifest folder. I'd like this to be smarter so that it just cancels the entire update function if it can't find the URL, but this needs to be in there for now. Deleted some temporary code that was commented out as well.

The manifest's location needs to be based on the source's name, but for now it's hardcoded until source configuration is supported. This will probably use a YAML file to store source names, locations, and types.