Move batch package installs to a closer...
version as I have an idea about how to implement batch package installs before the version that will have it go down the list in the Apply changes window. I still do want it to work like that eventually if the user wants to, but it may take more time.