Switch to LocalAppData instead of Roaming.
Not sure how much this will help, but it should help at least a little for networked profiles. There needs to be a thing that offers to delete the folder in Roaming for those who use the default updater. Maybe there can be a button on a "Refresh cache" tab in the Options window that deletes all files in AppData\Roaming\winget-frontends (possibly with the text "Delete old cache folder" with the label describing how it deletes the files in there and a link to open it if the user wants to; these controls should be deactivated if the folder doesn't exist) underneath a checkbox that deletes the temp files after the cache is refreshed.

See also #52.