Move database-updating code into its own If statement so it can...
work even if its temp folder already exists when the manifest temp folder also exists. This is a major step toward allowing people to update and use the database instead of just reading from the manifests. Actually, it's a major step toward allowing multiple sources based on a config file. Soon this will be pulled from a YAML file in Resources instead of being in code, though it may take a bit before it can read from a YAML file that's not in Resources. Hopefully soon there will be support for loading the Store database, but loading the manifests for that may take a bit more effort. Maybe that will require loading the remote manifest at runtime. Would prefer an easier method, but that list of manifests isn't available in a .zip folder as far as I know. At least each package has its own entry for the manifest path, so it should be relatively easy.

See also #35.