Allow deleting the current cache files from a button.
This was handled by just allowing the root path to be specified so it's easier to implement and can share code with the pre-existing button's code.

Probably should change this eventually so that there's a way to delete either the entire winget-frontends folder for clean uninstallation or just the cache in case of corruption.

Also added a message in case the directory seems to exist even after a successful deletion when the user is kicked out of the folder in Explorer.