Add export, import, and separator items to File + change the CMD...
stuff to just be CMD to save space. Also made the dialogs just say "Import Packages" as it's implied winget is being used for exporting, unless non-winget exporting is added eventually where you can export
a list of packages that are marked in the package list and import them for immediate marking.

I'm going to work on the importing and exporting stuff later as it's late.