Change the package details textbox to a RichTextBox so that...
the links can be clickable, though that code needs to be added. This has the side-effect of needing to get rid of the ".Replace(vbLf, vbCrLf)" stuff since it messed up a file that was using vbCrLf. This actually greatly improves the performance when displaying files so that you can just hold down an arrow key and go through them all in real-time. Another side-effect is right-clicking isn't supported anymore, but that wasn't really a feature, and you can use keyboard shortcuts instead.