Move "file path too long" message to its own sub...
outside the async block so that the boolean that says not to show it again this extraction gets updated properly. Haven't tested it yet, and this'll go into 0.4 once everything else is good rather than another update to the 0.3.x series as I'm tired of it straying so far off the main branch.