Make sure the tempPath exists before we delete...
it + add TODO regarding making sure the file doesn't exist yet before extracting, which can happen if we're not deleting the temp folder before beginning the extraction. There should be a way to only skip downloading a new zip file and database and still extract what's there for debugging purposes, instead of just loading from the moved manifests, but that isn't there yet.