We weren't properly forcing using the previous session's...
package list and I just noticed that by leaving the manifest repo zip file open in 7-Zip, so now we're using a boolean that's updated from a sub outside the Await, which is outside of the Await so it works properly. This may have ended up being a regression if I didn't know adding the Await would make the Try...Catch stop working.